feat: update scp and mv

+++ module_scp.rb
@@ -8,35 +8,33 @@ require 'zip'
 module SCP
     def compress(local_file, file_name)
         dest_dir = File.expand_path(local_file)
-        dest_file = File.join(dest_dir, '..', file_name)
+        dest_file = File.join(dest_dir, file_name)
         
         puts "#{local_file} ===> #{dest_dir} compress #{dest_file} ..."
 
-        if !File.exist?(dest_file)
-            puts "File does not exist."
+        if File.exist?(dest_file)
+            File.delete(dest_file) if File.exist?dest_file
+            puts 'Delete old version.'
         end
 
         puts "#{dest_file} loading..."
 
         begin
-            Zip::File.open(dest_file, Zip::File::CREATE) do |zipfile|
+            Zip::File.open(dest_file, create: true) do |zipfile|
+                zip_path = File.basename(dest_dir)
                 puts "entry #{zipfile}"
+
+                zipfile.mkdir(zip_path)
                 Dir.chdir(dest_dir) do
-                    Dir[File.join('**', '**')].each do |file|
+                    Dir[File.join('**', '**', '*')].each do |file|
+                        puts "files: #{file}"
                         if zipfile.find_entry(file)
                             puts "File already exists in the zip file: #{zipfile.find_entry(file)}, #{dest_file}" 
                         elsif zipfile.glob(file).any?
                             puts "File already exists in the zip file."
                         else
-                            # relative_path = file.sub(/^#{Regexp.escape(local_file + '/')}/, '')
-                            # puts "Path: #{relative_path}"
-                            # zipfile.add(relative_path, file)
-                            # zipfile.add(file.sub(local_file + "/", ''), file)
-                            if File.exist?(file)
-                                puts "Add: #{file}, #{file.sub(dest_dir + '/', '')}"
-                                # zipfile.add(file, file)
-                                zipfile.add(file.sub(dest_dir + '/', ''), file)
-                            end
+                            puts "Add: #{zip_path}/#{file}"
+                            zipfile.add(File.join(zip_path, file), "#{dest_dir}/#{file}")
                         end
                     end
                 end
@@ -50,12 +48,14 @@ module SCP
         return dest_file
     end
 
-    def move(file_path, target_path)
+    def move(local_file, target_path)
+        file_path = File.expand_path(local_file)
         FileUtils.mkdir_p(target_path) unless File.directory?(target_path)
 
         FileUtils.mv(file_path, target_path)
         
-        File.delete(file_path)
+        # delete local file
+        File.delete(file_path) if File.file?file_path
     end
 
     def unCompressfile(zip_file, target_dir)
@@ -63,13 +63,30 @@ module SCP
             zipfile.each do |file|
                 fpath = File.join(target_dir, file.name)
                 FileUtils.mkdir_p(File.dirname(fpath))
+
                 zipfile.extract(file, fpath) unless File.exist?(fpath)
             end
         end
 
         puts "Unzip successfully."
+
+        # delete remote file
         File.delete(zip_file)
         puts "Deleted #{zip_file} on remote machine."
+    end
+
+    def delete_file(local_file)
+        dir = File.expand_path(local_file)
+        puts "Do you really want to delete #{dir}? [Y/N]: "
+        answer = STDIN.gets.chomp.downcase
+
+        if answer == 'y'
+            print "exist: #{File.exist?dir}"
+            FileUtils.rm_rf(dir) if File.exist?dir
+            puts "Del #{dir} successfully!"
+        else
+            puts "Del canceled."
+        end
     end
 
     def upload_file(host, username, pwd, local_file, remote_file)
